refact: move interface
