+++ providers/aws/ec2/autoscaling_groups.go
@@ -15,6 +15,14 @@ import (
 
 const AWS_SERVICE_NAME_ASG = "AutoScalingGroup"
 
+type AutoScalingGroupClient interface {
+	DescribeAutoScalingGroups(
+		ctx context.Context,
+		params *autoscaling.DescribeAutoScalingGroupsInput,
+		optFns ...func(*autoscaling.Options),
+	) (*autoscaling.DescribeAutoScalingGroupsOutput, error)
+}
+
 func AutoScalingGroups(ctx context.Context, clientProvider ProviderClient) ([]Resource, error) {
 	client := autoscaling.NewFromConfig(*clientProvider.AWSClient)
 
@@ -36,7 +44,6 @@ type ASGDiscoverer struct {
 }
 
 func (d ASGDiscoverer) Discover() ([]Resource, error) {
-
 	resources := make([]Resource, 0)
 	var queryInput autoscaling.DescribeAutoScalingGroupsInput
